+++ libs/vault/src/services/vault-nudges/custom-nudges-services/index.ts
@@ -1,0 +1,1 @@
+export * from "./vault-onboarding-nudges.service";
