+++ libs/vault/src/services/vault-nudges/custom-nudges-services/vault-onboarding-nudges.service.ts
@@ -4,7 +4,8 @@ import { map, Observable, of, switchMap } from "rxjs";
 import { UserId } from "@bitwarden/common/types/guid";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 
-import { DefaultSingleNudgeService, VaultNudgeType } from "../vault-nudges.service";
+import { DefaultSingleNudgeService } from "../default-single-nudge.service";
+import { VaultNudgeType } from "../vault-nudges.service";
 
 /**
  * Custom Nudge Service to use for the Onboarding Nudges in the Vault
